Add validMoves value
Add an AngularJS value object containing all valid starting tiles as
properties and the valid progression as the values. If validMoves[m]
is defined the value is what the three tiles put in (i.e. the string
m) should be transformed into after the move is complete. If the
validMoves[m] is undefined, then m is an invalid point to do a move
from.