Fix bug in LevelCollection.collections()
The collections function on the LevelCollection service did not use
any kind of filtering, but only a map that produced a lot of
'undefined' values in the returned array of levels for each
collection. All collections had 355 levels, but most were undefined
except for the ones having the collection name currently traversed
in the levelData. Fix by filter, then map.