Build valid move generator to verify validMoves
In order to verify that the validMoves value in the app have states
for all valid moves we build a quick nodejs generator that generates
all valid starting states. This works by creating the dot product of
three arrays containing all valid states for the three tiles in
question. The obvious blockages are then removed.